requirements.txt
boto3==1.6.21
botocore==1.9.21
Django==1.11.11
django-filter==1.1.0
djangorestframework==3.7.7
docutils==0.14
futures==3.2.0
jmespath==0.9.3
Markdown==2.6.11
pkg-resources==0.0.0
python-dateutil==2.6.1
pytz==2018.3
s3transfer==0.1.13
six==1.11.0
